Render banner on static pages rendered as index
Normally, the banner is shown only on the index page generated from
the index direct template, or on all pages (if BANNER_ALL_PAGES is
set). We might want to render a static page as the index though, or a
page rendered from a different template, that is turned into the main
index page via a save_as attribute.

Add an additional test that also renders the banner if output_file is
'index.html'.